cql3: move prepared_cache_key_type to a lightweight header
Extract prepared_cache_key_type and its hash and fmt specializations from
prepared_statements_cache.hh into cql3/prepared_cache_key_type.hh.

This keeps the cache key type available without pulling in the loading
cache implementation header.
